seperate ide into files
