+++ src/cogs/ide/dialogs/__init__.py
@@ -1,0 +1,3 @@
+from .OpenView import OpenView
+from .FileView import FileView
+from .EditView import EditView
