+++ src/utils/__init__.py
@@ -1,2 +1,3 @@
 from .paginator import TextPaginator
 from .confirmation import prompt
+from .utils import *
